refactor(all): separate files into folders
Closes #37
