+++ lib/main.dart
@@ -1,5 +1,5 @@
 
-import 'package:easymakers_tracker/home_page.dart';
+import 'package:easymakers_tracker/ui/pages/home_page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
