+++ lib/ui/pages/clients_page.dart
@@ -1,6 +1,6 @@
-import 'package:easymakers_tracker/client.dart';
-import 'package:easymakers_tracker/client_form.dart';
-import 'package:easymakers_tracker/client_storage.dart';
+import 'package:easymakers_tracker/models/client.dart';
+import 'package:easymakers_tracker/ui/forms/client_form.dart';
+import 'package:easymakers_tracker/stores/client_storage.dart';
 import 'package:flutter/material.dart';
 
 class ClientsPage extends StatefulWidget {
