+++ test/home_page_test.dart
@@ -1,5 +1,5 @@
-import 'package:easymakers_tracker/easymakers_page.dart';
-import 'package:easymakers_tracker/home_page.dart';
+import 'package:easymakers_tracker/ui/pages/easymakers_page.dart';
+import 'package:easymakers_tracker/ui/pages/home_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
