+++ test/mission_form_test.dart
@@ -1,6 +1,6 @@
-import 'package:easymakers_tracker/client.dart';
-import 'package:easymakers_tracker/easymaker.dart';
-import 'package:easymakers_tracker/mission_form.dart';
+import 'package:easymakers_tracker/models/client.dart';
+import 'package:easymakers_tracker/models/easymaker.dart';
+import 'package:easymakers_tracker/ui/forms/mission_form.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
